feat : 매칭 내역 -> 매칭 명세표 -> 프로필 #40

+++ app/src/main/java/com/example/fit_i/data/model/response/GetMatchlistResponse.kt
@@ -7,8 +7,8 @@ data class GetMatchlistResponse(
     val result : Result
 ){
     data class Result(
-        val matchingId : Int,
-        val customerId : Int,
+        val matchingId : Long,
+        val customerId : Long,
         val name : String,
         val pricePerHour : String,
         val totalPrice : String,
